refactor: move pure data ops from cmd/doc to ops/doc
Move parse_value, flatten_value, diff_values, and DiffEntry from
src/cmd/doc.rs to src/ops/doc/mod.rs. These are pure data operations
with no CLI dependencies (no GlobalFlags, no clap, no exit codes),
so they belong in the ops layer.

This improves layering: the ops layer provides reusable data
primitives, and the cmd layer handles CLI concerns (argument parsing,
output formatting, exit codes).

Changes:
- ops/doc/mod.rs: add parse_value, flatten_value, DiffEntry,
  diff_values as pub functions with tests
- cmd/doc.rs: remove local definitions, import from ops::doc
- cmd/batch.rs: update parse_value import path

Net: -188 lines from cmd/doc.rs, +238 lines in ops/doc/mod.rs
(includes new unit tests for the moved functions)
